Fix constant paint requests from CrocAnalogMeter
The analog meter was calling setText on it's label which then requests a repaint, meaning that the analog meter would constantly be looping repaint request all the time. Probably will need to put in some check for the CrocRoot to not accept a paint request during painting, or to maybe just change when we set the dirty flag for repaints to after the painting is done and not before.